api/requirements.txt
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2